Fixed cleanup methods

+++ compute/cloud-client/src/test/java/compute/Util.java
@@ -166,6 +166,9 @@ public abstract class Util {
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
     try (ReservationsClient reservationsClient = ReservationsClient.create()) {
       for (Reservation reservation : reservationsClient.list(projectId, zone).iterateAll()) {
+        if (!reservationsClient.list(projectId, zone).iterateAll().iterator().hasNext()) {
+          continue;
+        }
         if (reservation.getName().contains(prefixToDelete)) {
           DeleteReservation.deleteReservation(projectId, zone, reservation.getName());
         }
@@ -178,6 +181,9 @@ public abstract class Util {
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
     try (DisksClient disksClient = DisksClient.create()) {
       for (Disk disk : disksClient.list(projectId, zone).iterateAll()) {
+        if (!disksClient.list(projectId, zone).iterateAll().iterator().hasNext()) {
+          continue;
+        }
         if (disk.getName().contains(prefixToDelete)) {
           DeleteDisk.deleteDisk(projectId, zone, disk.getName());
         }
@@ -189,6 +195,9 @@ public abstract class Util {
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
     try (SnapshotsClient snapshotsClient = SnapshotsClient.create()) {
       for (Snapshot snapshot : snapshotsClient.list(projectId).iterateAll()) {
+        if (!snapshotsClient.list(projectId).iterateAll().iterator().hasNext()) {
+          continue;
+        }
         if (snapshot.getName().contains(prefixToDelete)) {
           DeleteSnapshot.deleteSnapshot(projectId, snapshot.getName());
         }
